fix: update GetStreamId method in SmuxParam interface to use int32 for improved compatibility

+++ example/message/message.go
@@ -25,7 +25,7 @@ type Packet struct {
 	Index       int32  `json:"index"`
 	Data        []byte `json:"data"`
 	DataVersion uint64 `json:"data_version"`
-	StreamID    int64  `json:"stream_id"`
+	StreamID    int32  `json:"stream_id"`
 	FragID      int32  `json:"frag_id"`
 	FragCount   int32  `json:"frag_count"`
 	FragIndex   int32  `json:"frag_index"`
@@ -66,7 +66,7 @@ func (m *Packet) GetDataVersion() uint64 {
 	return m.DataVersion
 }
 
-func (m *Packet) GetStreamId() int64 {
+func (m *Packet) GetStreamId() int32 {
 	return m.StreamID
 }
